Include functional options in main file
