+++ manifestival.go
@@ -42,6 +42,23 @@ type Manifest struct {
 	log       logr.Logger
 }
 
+// Option enables configuration of your Manifest
+type Option func(*Manifest)
+
+// UseLogger will cause manifestival to log its actions
+func UseLogger(log logr.Logger) Option {
+	return func(m *Manifest) {
+		m.log = log
+	}
+}
+
+// UseClient enables interaction with the k8s API server
+func UseClient(client client.Client) Option {
+	return func(m *Manifest) {
+		m.Client = client
+	}
+}
+
 var _ Manifestival = &Manifest{}
 
 // NewManifest creates a Manifest from a comma-separated set of YAML
